chore(dragdrop): refactor styles
